feat: change vector tile get data
