+++ dev-demos/layer/control/mapTheme.tsx
@@ -1,11 +1,9 @@
 import { GaodeMap, Scene, MapTheme } from '@antv/l7';
-import React, { useState } from 'react';
+import React from 'react';
 // tslint:disable-next-line:no-duplicate-imports
 import { FunctionComponent, useEffect } from 'react';
 
 const Demo: FunctionComponent = () => {
-  const [scene, setScene] = useState<Scene | undefined>();
-
   useEffect(() => {
     const newScene = new Scene({
       id: 'map',
