+++ packages/component/src/control/fullscreen.ts
@@ -47,6 +47,7 @@ export default class Fullscreen extends ButtonControl<
   public onAdd(): HTMLElement {
     const button = super.onAdd();
     button.addEventListener('click', this.onClick);
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
     this.mapContainer = DOM.getContainer(this.scene.getSceneConfig().id!);
     this.mapContainer.addEventListener(
       'fullscreenchange',
