+++ packages/source/src/parser/mvt.ts
@@ -6,12 +6,8 @@ import {
   TilesetManagerOptions,
   RequestParameters,
 } from '@antv/l7-utils';
-import {
-  VectorTile,
-  // VectorTileFeature,
-  VectorTileLayer,
-} from '@mapbox/vector-tile';
-import { Feature, Properties } from '@turf/helpers';
+import { VectorTile, VectorTileLayer } from '@mapbox/vector-tile';
+import { Feature } from '@turf/helpers';
 import Protobuf from 'pbf';
 import { IParserData } from '../interface';
 import { ITileParserCFG } from '@antv/l7-core';
@@ -24,15 +20,6 @@ const DEFAULT_CONFIG: Partial<TilesetManagerOptions> = {
   warp: true,
 };
 
-// const TILE_SIZE = 512;
-
-export function osmTileXY2LonLat(x: number, y: number, zoom: number) {
-  const lon = (x / Math.pow(2, zoom)) * 360 - 180;
-  const n = Math.PI - (2 * Math.PI * y) / Math.pow(2, zoom);
-  const lat = (180 / Math.PI) * Math.atan(0.5 * (Math.exp(n) - Math.exp(-n)));
-  return [lon, lat];
-}
-
 export type MapboxVectorTile = {
   layers: { [_: string]: VectorTileLayer & { features: Feature[] } };
 };
@@ -42,7 +29,6 @@ const getVectorTile = async (
   tileParams: TileLoadParams,
   tile: SourceTile,
   requestParameters?: Partial<RequestParameters>,
-  // coord: string,
 ): Promise<MapboxVectorTile> => {
   const tileUrl = getURLFromTemplate(url, tileParams);
   return new Promise((resolve) => {
@@ -53,40 +39,11 @@ const getVectorTile = async (
       },
       (err, data) => {
         if (err || !data) {
-          // reject(err);
           resolve({ layers: {} });
         } else {
           const vectorTile = new VectorTile(
             new Protobuf(data),
           ) as MapboxVectorTile;
-          // check tile source layer
-          for (const sourceLayer of Object.keys(vectorTile.layers)) {
-            const features: Feature<GeoJSON.Geometry, Properties>[] = [];
-            const vectorTileLayer = vectorTile.layers[sourceLayer];
-            for (let i = 0; i < vectorTile.layers[sourceLayer].length; i++) {
-              const vectorTileFeature = vectorTile.layers[sourceLayer].feature(
-                i,
-              );
-              // let feature;
-              // if (coord === 'lnglat') {
-              const feature = vectorTileFeature.toGeoJSON(
-                tileParams.x,
-                tileParams.y,
-                tileParams.z,
-              );
-              // TODO ID 统一编码
-              features.push({
-                ...feature,
-                properties: {
-                  id: feature.id,
-                  ...feature.properties,
-                },
-              });
-            }
-            // @ts-ignore
-            vectorTileLayer.features = features;
-          }
-
           resolve(vectorTile);
         }
       },
@@ -102,10 +59,8 @@ export default function mapboxVectorTile(
   // TODO: 后续考虑支持多服务
   const url = Array.isArray(data) ? data[0] : data;
 
-  // const coord = cfg?.coord || 'lnglat'; // lnglat - offset
   const getTileData = (tileParams: TileLoadParams, tile: SourceTile) =>
     getVectorTile(url, tileParams, tile, cfg?.requestParameters);
-  // getVectorTile(data, tileParams, tile, coord);
 
   const tilesetOptions = {
     ...DEFAULT_CONFIG,
